Feat(order): 주문 포인트 타입 BigDecimal -> Long

+++ order-service/src/main/java/com/rushcrew/order_service/application/command/dto/command/CreateOrderCommand.java
@@ -16,7 +16,7 @@ public record CreateOrderCommand(
 	String queueToken,
 	String role,
 	List<OrderItemCommand> orderItems,
-	BigDecimal pointUsed,
+	Long pointUsed,
 	ShippingInfo shippingInfo
 ) {
