+++ order-service/src/main/java/com/rushcrew/order_service/application/command/dto/result/CreateOrderResult.java
@@ -13,7 +13,7 @@ public record CreateOrderResult(
 	Long userId,
 	String orderStatus,
 	BigDecimal totalAmount,
-	BigDecimal pointUsed,
+	Long pointUsed,
 	BigDecimal finalAmount,
 	Instant orderedAt,
 	Instant reservationExpiresAt,	// 예약 만료 시간: 15분 후
